requirements.txt
Flask==3.0.0
pymongo==4.6.0
python-dateutil==2.8.2
requests==2.31.0

